Add 'cicustom' input variable for custom cloudinit
This adds a `cicustom` input variable for specifying a custom cloud-init
configuration.

This variable is mutually exclusive with `admin_username`,
`admin_password`, and `admin_public_ssh_keys`. If `cicustom` is set,
those variables should be left unset to `null`. The `admin_username`
variable now defaults to `null`.

The handling of the user data itself is not handled in this module and a
path to a cloud-init config on Proxmox storage should be set.

The syntax for a value looks like
`user=local:snippets/user-data_vm-foo.yml`, where `local` is the storage
name, `snippets` is the path on the filesystem (`/var/lib/vz/snippets`),
and `user-data_vm-foo.yml` is the user data config file deployed
external to this module.