Truncate /etc/machine-id with any changes made
When running this as-is, `virt-customize` bakes an /etc/machine-id into the Ubuntu image which is very problematic for a template application. Each machine should have a unique ID, otherwise MAC addresses (and, therefore, DHCP IP addressing) will be duplicated in the network. Appending `--truncate /etc/machine-d` removes the machine ID that this script creates when it installs the package and/or runs any other scripts.